requirement.txt
python 3.9
pandas
sklearn
sklearn
textstat
pyenchant
transformers
pytorch : conda install pytorch torchvision torchaudio pytorch-cuda=11.8 -c pytorch -c nvidia
